2_dataset_generation/requirements.txt
# Google Earth Engine
earthengine-api>=0.1.350
google-auth>=2.16.0

# Data manipulation
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Configuration
pyyaml>=6.0

# Metadata storage (Parquet format)
pyarrow>=10.0.0

# Visualization
matplotlib>=3.5.0